Renamed emails to match what they actually do
