+++ packages/emails/src/templates/AttendeeRescheduledEmail.tsx
@@ -4,7 +4,7 @@ export const AttendeeRescheduledEmail = (props: React.ComponentProps<typeof Atte
   <AttendeeScheduledEmail
     title="event_has_been_rescheduled"
     headerType="calendarCircle"
-    subject="rescheduled_event_type_subject"
+    subject="event_type_has_been_rescheduled_on_time_date"
     {...props}
   />
 );
