+++ packages/emails/src/templates/OrganizerRescheduledEmail.tsx
@@ -4,7 +4,7 @@ export const OrganizerRescheduledEmail = (props: React.ComponentProps<typeof Org
   <OrganizerScheduledEmail
     title="event_has_been_rescheduled"
     headerType="calendarCircle"
-    subject="rescheduled_event_type_subject"
+    subject="event_type_has_been_rescheduled_on_time_date"
     {...props}
   />
 );
