+++ packages/emails/src/templates/index.ts
@@ -3,7 +3,7 @@ export { AttendeeCancelledEmail } from "./AttendeeCancelledEmail";
 export { AttendeeDeclinedEmail } from "./AttendeeDeclinedEmail";
 export { AttendeeLocationChangeEmail } from "./AttendeeLocationChangeEmail";
 export { AttendeeRequestEmail } from "./AttendeeRequestEmail";
-export { AttendeeRequestRescheduledEmail } from "./AttendeeRequestRescheduledEmail";
+export { AttendeeWasRequestedToRescheduleEmail } from "./AttendeeWasRequestedToRescheduleEmail";
 export { AttendeeRescheduledEmail } from "./AttendeeRescheduledEmail";
 export { AttendeeScheduledEmail } from "./AttendeeScheduledEmail";
 export { FeedbackEmail } from "./FeedbackEmail";
@@ -13,7 +13,7 @@ export { OrganizerLocationChangeEmail } from "./OrganizerLocationChangeEmail";
 export { OrganizerPaymentRefundFailedEmail } from "./OrganizerPaymentRefundFailedEmail";
 export { OrganizerRequestEmail } from "./OrganizerRequestEmail";
 export { OrganizerRequestReminderEmail } from "./OrganizerRequestReminderEmail";
-export { OrganizerRequestRescheduledEmail } from "./OrganizerRequestRescheduledEmail";
+export { OrganizerRequestedToRescheduleEmail } from "./OrganizerRequestedToRescheduleEmail";
 export { OrganizerRescheduledEmail } from "./OrganizerRescheduledEmail";
 export { OrganizerScheduledEmail } from "./OrganizerScheduledEmail";
 export { TeamInviteEmail } from "./TeamInviteEmail";
