+++ packages/emails/templates/attendee-rescheduled-email.ts
@@ -11,12 +11,12 @@ export default class AttendeeRescheduledEmail extends AttendeeScheduledEmail {
       to: `${this.attendee.name} <${this.attendee.email}>`,
       from: `${this.calEvent.organizer.name} <${this.getMailerOptions().from}>`,
       replyTo: [...this.calEvent.attendees.map(({ email }) => email), this.calEvent.organizer.email],
-      subject: `${this.attendee.language.translate("rescheduled_event_type_subject", {
+      subject: `${this.attendee.language.translate("event_type_has_been_rescheduled_on_time_date", {
         eventType: this.calEvent.type,
         name: this.calEvent.team?.name || this.calEvent.organizer.name,
         date: this.getFormattedDate(),
       })}`,
-      html: renderEmail("AttendeeCancelledEmail", {
+      html: renderEmail("AttendeeRescheduledEmail", {
         calEvent: this.calEvent,
         attendee: this.attendee,
       }),
